Fix gate by bumping tox version and removing ryu
Gating jobs started to fail with latest pyproject-api (1.6.1) release.
This release would need latest tox (4.11.0) to not fail, but we are
pinning tox via pinning virtualenv in tox.ini.
To resolve this we have to remove the virtualenv pinning, which was a
"temporary" hack to avoid py3*-check-uc jobs to break. This job is
failing because of ryu package, which is unmaintained and projects
started to use os_ken instead, which "is a fork of the Ryu library
tailored for OpenStack Neutron" [1]. Hence, this patch removes ryu
from requirements files to eliminate the need of pinning virtualenv.

[1] https://opendev.org/openstack/os-ken

Change-Id: I23ba892e6dd7f2566e5af6626c072c2dcc5a4382
(cherry picked from commit b3ccca8516377c16663217efe20efa1593d74727)
Signed-off-by: Mohammed Naser <mnaser@vexxhost.com>